Renamed applicant models and controllers to application
