+++ app/controllers/application_pet_controller.rb
@@ -1,0 +1,3 @@
+class ApplicationPetController < ActionController
+  
+end
