+++ app/models/application.rb
@@ -1,0 +1,8 @@
+class Application < ApplicationRecord
+  has_many :application_pets
+  has_many :pets, through: :application_pets
+
+  def full_address 
+    "#{street_address}, #{city}, #{state} #{zip_code}"
+  end
+end
