+++ app/views/applications/new.html.erb
