+++ app/views/applications/show.html.erb
