+++ spec/features/applications/show_spec.rb
